refactor: 좋아요 svg 이미지 사용 방식 변경
