+++ src/components/ProductCard/ProductCard.jsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import { css } from "@emotion/react";
 import { Link } from "react-router-dom";
-import likeImg from "../../assets/images/ic-like.svg";
+import { ReactComponent as HeartIcon } from "@/assets/images/ic-like.svg";
 import pandaLogoImg from "../../assets/images/logo-panda.svg";
 import { getProduct } from "@/services/get/getProduct";
 import useAsync from "@/hooks/useAsync";
@@ -49,7 +49,12 @@ const ProductCard = ({ productId, data, loading = "lazy" }) => {
         <h4 className="item-price">{price.toLocaleString("ko-KR")}원</h4>
         <button className="btn-like">
           <span className="btn-like-ico">
-            <img className="ico-img" src={likeImg} alt="좋아요" />
+            <HeartIcon
+              aria-label="좋아요 갯수"
+              className="heart-icon"
+              width="100%"
+              height="100%"
+            />
           </span>
           <span className="btn-like-count">{favoriteCount}</span>
         </button>
@@ -95,8 +100,10 @@ export const ProductCardStyle = css`
     height: 16px;
   }
 
-  .btn-like-ico .ico-img {
+  .btn-like-ico .heart-icon {
     height: 100%;
     object-fit: contain;
+    stroke: 1px solid var(--gray600);
+    fill: #fff;
   }
 `;
